[refactor] Comment api변경에 따른 디렉토리 구조 변경
