+++ backend/routes/api/issues/[no]/index.js
@@ -5,7 +5,6 @@ import labelRouter from './labels';
 import milestoneRouter from './milestone';
 import openRouter from './open';
 import closeRouter from './close';
-import commentRouter from './comments';
 import * as issueNoController from './no.controller';
 
 const router = express.Router({ mergeParams: true });
@@ -14,7 +13,6 @@ router.use('/title', titleRouter);
 router.use('/assignees', assigneeRouter);
 router.use('/labels', labelRouter);
 router.use('/milestone', milestoneRouter);
-router.use('/comments', commentRouter);
 router.use('/open', openRouter);
 router.use('/close', closeRouter);
